04-analytics-engineering/requirements.txt
# Module 4 - Analytics Engineering with dbt

# LOCAL DEVELOPMENT (DuckDB):
#   pip install -r requirements.txt

# SWITCHING BETWEEN ENVIRONMENTS:
#   dbt run --target dev   (uses DuckDB)
#   dbt run --target prod  (uses BigQuery)


# Python dependencies

# dbt with DuckDB adapter (for local development)
dbt-core==1.7.4
dbt-duckdb==1.7.2

# Data manipulation
pandas==2.1.4
pyarrow==14.0.2

# Utilities for data download
requests==2.31.0
tqdm==4.66.1

# Optional: Data quality
great-expectations==0.18.8

# Optional: BigQuery adapter (for production deployment)
# dbt-bigquery==1.7.4
# google-cloud-bigquery==3.14.1

protobuf<5.0.0,>=4.0.0

# Table formatting (optional, for pretty output)
tabulate==0.9.0

# LOCAL DEVELOPMENT (DuckDB):
#   pip install -r requirements.txt
#   Uses local parquet files, no GCP needed

# PRODUCTION (BigQuery):
#   1. Uncomment dbt-bigquery lines above
#   2. pip install -r requirements.txt
#   3. Configure profiles.yml with your GCP project
#   4. Authenticate: gcloud auth application-default login


# SWITCHING BETWEEN ENVIRONMENTS:
#   dbt run --target dev   (uses DuckDB)
#   dbt run --target prod  (uses BigQuery)